feat: use cytoscape for graph and cleanup
