+++ frontend/console/src/features/graph/ExpandablePanel.tsx
