+++ frontend/console/src/features/graph/ModulePanels.tsx
@@ -2,9 +2,9 @@ import { ArrowLeftRightIcon, FunctionIcon, InboxDownloadIcon, InboxUploadIcon, L
 import type { NavigateFunction } from 'react-router-dom'
 import { RightPanelAttribute } from '../../components/RightPanelAttribute'
 import type { Module } from '../../protos/xyz/block/ftl/v1/console/console_pb'
+import type { ExpandablePanelProps } from '../graph/ExpandablePanel'
 import { callsIn, callsOut } from '../modules/module.utils'
 import { Schema } from '../modules/schema/Schema'
-import type { ExpandablePanelProps } from './ExpandablePanel'
 
 export const modulePanels = (allModules: Module[], module: Module, navigate: NavigateFunction): ExpandablePanelProps[] => {
   const panels = []
