+++ frontend/console/src/features/modules/decls/DeclDefaultPanels.tsx
@@ -1,5 +1,5 @@
 import type { Ref } from '../../../protos/xyz/block/ftl/v1/schema/schema_pb'
-import type { ExpandablePanelProps } from '../../console/ExpandablePanel'
+import type { ExpandablePanelProps } from '../../graph/ExpandablePanel'
 import { Schema } from '../schema/Schema'
 import { References } from './References'
