+++ frontend/console/src/features/modules/decls/verb/VerbRightPanel.tsx
@@ -2,7 +2,7 @@ import { InboxUploadIcon } from 'hugeicons-react'
 import { Link } from 'react-router-dom'
 import { RightPanelAttribute } from '../../../../components/RightPanelAttribute'
 import type { Verb } from '../../../../protos/xyz/block/ftl/v1/console/console_pb'
-import type { ExpandablePanelProps } from '../../../console/ExpandablePanel'
+import type { ExpandablePanelProps } from '../../../graph/ExpandablePanel'
 import { Schema } from '../../schema/Schema'
 import { type VerbRef, httpRequestPath, ingress, isHttpIngress, verbCalls } from './verb.utils'
